SEQUOIADBMAINSTREAM-6563: fix arm64 align issue

+++ SequoiaDB/engine/opt/optPlanNode.cpp
@@ -57,7 +57,7 @@ using namespace bson;
 // 2. By using user specified allocator(instances of optPlanAllocator).
 // The actions when releasing these two kinds of node are different.
 // So a type flag is added at the head of the actual allocated memory.
-#define OPT_MEM_TYPE_SIZE            sizeof(INT32)
+#define OPT_MEM_TYPE_SIZE            sizeof(INT64)
 #define OPT_MEM_BY_USER_ALLOCATOR    0
 #define OPT_MEM_BY_DFT_ALLOCATOR     1
